Reorganize Plugin Registry so it has a db/, nomad/, server/
