+++ src/rust/plugin-registry/src/db/client.rs
