+++ src/rust/plugin-registry/src/nomad/cli.rs
@@ -17,6 +17,7 @@ pub enum NomadCliError {
 
 pub type NomadVars = HashMap<String, String>;
 
+#[derive(Default)]
 pub struct NomadCli {}
 
 impl NomadCli {
