+++ src/rust/plugin-registry/src/nomad/client.rs
