+++ src/rust/plugin-registry/src/nomad/mod.rs
@@ -1,0 +1,2 @@
+pub mod cli;
+pub mod client;
