+++ src/rust/plugin-registry/src/server/mod.rs
@@ -1,0 +1,3 @@
+mod deploy_plugin;
+pub mod main;
+mod server;
