+++ src/rust/plugin-registry/tests/test_nomad_client.rs
@@ -1,10 +1,10 @@
 #![cfg(feature = "integration")]
 
-use plugin_registry::nomad_client;
+use plugin_registry::nomad::client::NomadClient;
 
 #[test_log::test(tokio::test)]
 async fn test_nomad_client_create_namespace() -> Result<(), Box<dyn std::error::Error>> {
-    let client = nomad_client::NomadClient::from_env();
+    let client = NomadClient::from_env();
     client
         .create_namespace("test-nomad-client-create-namespace")
         .await?;
